Prepare ruff-python-ast crate
